backend/requirements.txt
fastapi
uvicorn
requests
google-generativeai
python-dotenv
python-multipart
openai-whisper
beautifulsoup4
sentence-transformers
scikit-learn
routers
pillow